Move execution Context into instruction-executor
