+++ packages/machine/src/middleware/state-transition/update-proposer.ts
@@ -1,6 +1,7 @@
 import * as cf from "@counterfactual/cf.js";
 
-import { Context, NodeState } from "../../node-state";
+import { Context } from "../../instruction-executor";
+import { NodeState } from "../../node-state";
 import { InternalMessage, StateProposal } from "../../types";
 
 export class UpdateProposer {
