+++ packages/machine/src/write-ahead-log.ts
@@ -1,6 +1,6 @@
 import * as cf from "@counterfactual/cf.js";
 
-import { Context } from "./node-state";
+import { Context } from "./instruction-executor";
 import { InternalMessage, MiddlewareResult } from "./types";
 
 /**
